client - validation moved to src/modules/common/Validator.js
