+++ client/src/modules/common/Button.js
@@ -1,9 +1,6 @@
 import React from 'react';
 
 class Button extends React.Component {
-    constructor(props) {
-        super(props);
-    }
 
     render() {
         return (
@@ -17,4 +14,5 @@ class Button extends React.Component {
         );
     }
 }
+
 export default Button;
